refactor: relocate store and add RootState & AppDispatch type declarations
