+++ src/app/providers/StoreProvider.tsx
@@ -1,7 +1,7 @@
 'use client'
-import { makeStore } from '@/shared/config'
 import { type ReactNode } from 'react'
 import { Provider } from 'react-redux'
+import { makeStore } from '../store'
 
 export const StoreProvider = ({ children }: { children: ReactNode }) => {
   const store = makeStore()
