+++ src/app/store.ts
@@ -15,8 +15,3 @@ export const makeStore = () => {
 }
 
 setupListeners(makeStore().dispatch)
-
-export type AppStore = ReturnType<typeof makeStore>
-// Infer the `RootState` and `AppDispatch` types from the store itself
-export type RootState = ReturnType<AppStore['getState']>
-export type AppDispatch = AppStore['dispatch']
