+++ src/app/types/index.d.ts
@@ -1,0 +1,2 @@
+declare type RootState = ReturnType<ReturnType<typeof import('../store')['makeStore']>['getState']>;
+declare type AppDispatch = ReturnType<typeof import('../store')['makeStore']>['dispatch'];
