+++ src/shared/config/index.ts
@@ -1,3 +1,2 @@
 export * from './backend'
-export * from './store'
 export * from './i18n'
