+++ src/shared/lib/hooks/index.ts
@@ -1,3 +1,1 @@
-export { useAppDispatch } from './useAppDispatch'
-export { createAppAsyncThunk } from './createAppAsyncThunk'
 export { getErrorMessage, getStatusCode } from './getErrorMessage'
