--- src/shared/lib/hooks/useAppDispatch.ts
@@ -1,4 +1,0 @@
-import type { AppDispatch } from '@/shared/config'
-import { useDispatch } from 'react-redux'
-
-export const useAppDispatch = () => useDispatch<AppDispatch>()
