+++ src/shared/lib/store/index.ts
@@ -1,0 +1,1 @@
+export { createAppAsyncThunk, useAppDispatch } from './redux'
