+++ src/shared/lib/store/redux.ts
@@ -1,5 +1,7 @@
-import type { AppDispatch, RootState } from '@/shared/config'
 import { createAsyncThunk } from '@reduxjs/toolkit'
+import { useDispatch } from 'react-redux'
+
+export const useAppDispatch = () => useDispatch<AppDispatch>()
 
 export const createAppAsyncThunk = createAsyncThunk.withTypes<{
   state: RootState
